jQuery, Handlebars and Prism from npm
