+++ design-system/package.json
@@ -51,7 +51,10 @@
     "@coopdigital/shared-component--statement": "^3.1.1",
     "@coopdigital/shared-component--totaliser": "^1.1.1",
     "@coopdigital/shared-component--video": "^2.2.1",
-    "normalize.css": "^8.0.1"
+    "handlebars": "^4.7.6",
+    "jquery": "^3.5.1",
+    "normalize.css": "^8.0.1",
+    "prismjs": "^1.23.0"
   },
   "scripts": {
     "build": "gulp build",
@@ -59,10 +62,17 @@
     "server": "gulp server"
   },
   "devDependencies": {
+    "@babel/core": "^7.12.13",
+    "@babel/plugin-transform-runtime": "^7.12.15",
+    "@babel/preset-env": "^7.12.13",
     "autoprefixer": "^10.0.2",
+    "babelify": "^10.0.0",
+    "browserify": "^17.0.0",
+    "core-js": "^3.8.3",
     "cssnano": "^4.1.10",
     "gulp": "^4.0.2",
     "gulp-autoprefixer": "^7.0.1",
+    "gulp-buffer": "0.0.2",
     "gulp-changed": "^4.0.2",
     "gulp-concat": "^2.6.1",
     "gulp-connect": "^5.7.0",
@@ -70,7 +80,8 @@
     "gulp-include": "^2.4.1",
     "gulp-postcss": "^9.0.0",
     "gulp-sourcemaps": "^3.0.0",
-    "gulp-uglify": "^3.0.2",
+    "gulp-tap": "^2.0.0",
+    "gulp-terser": "^2.0.1",
     "postcss": "^8.2.4",
     "postcss-calc": "^8.0.0",
     "postcss-custom-media": "^8.0.0",
