+++ design-system/src/_js/includes/coop-toolkit.js
@@ -1,14 +1,17 @@
+const $ = require('jquery');
+
 // Adapted from https://github.com/Heisenbergjs/heisenberg/blob/master/app/app.js
-(function() {
+/* eslint-disable */
+(function () {
   window.Coop = window.Coop || {
 
     Modules: {},
 
     Events: $({}),
 
-    init: function () {
+    init() {
       // Loop through all the modules added to the coop.Modules object. We check that the init function of each Module can be called.
-      for(var x in Coop.Modules) {
+      for (const x in Coop.Modules) {
         if (Coop.Modules[x].ModuleInitDone !== true && typeof Coop.Modules[x].init === 'function') {
           Coop.Modules[x].init();
           Coop.Modules[x].ModuleInitDone = true;
@@ -16,6 +19,7 @@
       }
 
       Coop.Events.trigger('render');
-    }
+    },
   };
-})();
+}());
+/* eslint-enable */
