+++ design-system/src/_js/includes/syntax-highlighting.js
@@ -1,0 +1,8 @@
+window.Prism = require('prismjs');
+
+// Syntax highlighting languages
+require('prismjs/components/prism-javascript.js');
+require('prismjs/components/prism-markup.js');
+require('prismjs/components/prism-scss.js');
+
+window.Prism.highlightAll();
