+++ design-system/src/_layouts/default.html
@@ -26,10 +26,6 @@
         </main>
     </div>
     {% include feedback.html %}
-    <script src="https://code.jquery.com/jquery-1.12.1.min.js" integrity="sha256-I1nTg78tSrZev3kjvfdM5A5Ak/blglGzlaZANLPDl3I=" crossorigin="anonymous"></script>
-    <script>window.jQuery || document.write('<script src="/assets/js/vendor/jquery-1.12.1.min.js"><\/script>')</script>
-    <script src="https://cdn.rawgit.com/filamentgroup/fixed-sticky/master/fixedsticky.js" crossorigin="anonymous"></script>
     <script src="/assets/js/main.js?v={{ site.time | date: '%s' }}"></script>
-    <script src="/assets/js/vendor/prism.js?v={{ site.time | date: '%s' }}"></script>
 </body>
 </html>
